+++ NoteManager/src/test/java/it/libero/alessandragenca/notesmanager/server/web/resources/NoteRegSizeJSONTest.java
@@ -46,7 +46,7 @@ public class NoteRegSizeJSONTest {
         try
         {
 
-            Scanner scanner = new Scanner(new File("src/main/resources/settings.json"));
+            Scanner scanner = new Scanner(new File("src/main/resources/settings.json"),"UTF-8");
             settings = gson.fromJson(scanner.nextLine(),NoteRegSizeJSONTest.Settings.class);
             scanner.close();
             //System.err.println("Loading settings from file");
@@ -156,10 +156,14 @@ public class NoteRegSizeJSONTest {
     public static void tearDownAfterClass()  {
         File file = new File("src/main/resources/storage");
         if (file.exists()) {
-            for (File f : file.listFiles()) {
-                //System.out.println(f.getName());
-                if (f.getName().startsWith("Test")) {
-                    f.deleteOnExit();
+            File[] listFiles=file.listFiles();
+
+            if(listFiles!=null){
+                for (File f : listFiles) {
+                    //System.out.println(f.getName());
+                    if (f.getName().startsWith("Test")) {
+                        f.deleteOnExit();
+                    }
                 }
             }
         }
